# [Impeller] Joined obligatory vulkan swapchain submits