@gaaclarke 2023-06-16
Here is the skiaperf for this PR:
https://flutter-flutter-perf.skia.org/e/?end=1687038803&keys=Xc266cefa78856177cd92ded480267478&selected=commit%3D35358%26name%3D%252Carch%253Dintel%252Cbranch%253Dmaster%252Cconfig%253Ddefault%252Cdevice_type%253DSM-G973U1%252Cdevice_version%253Dnone%252Chost_type%253Dlinux%252Csub_result%253Daverage_frame_rasterizer_time_millis%252Ctest%253Dnew_gallery_impeller__transition_perf%252C

It doesn't show any decrease in the average frame time.  What I've learned is:
1) That benchmark is noisy.  Quite frequently it will spike from ~10ms to ~20ms for an average over ~350 samples.  I suspect that was related to the decreased measured above
1) My local tests above were doing performance benchmarks with validation layers turned on, so that may have played a roll in the reduction in time too, but is inconsequential for end users

I ran the tests again 3 times locally:

## before
```
"average_frame_build_time_millis": 1.6772485875706202,
  "90th_percentile_frame_build_time_millis": 3.643,
  "99th_percentile_frame_build_time_millis": 17.397,
  "worst_frame_build_time_millis": 31.351,
  "missed_frame_build_budget_count": 8,
  "average_frame_rasterizer_time_millis": 9.822338983050852,
  "90th_percentile_frame_rasterizer_time_millis": 12.731,
  "99th_percentile_frame_rasterizer_time_millis": 98.295,
  "worst_frame_rasterizer_time_millis": 109.299,
  "missed_frame_rasterizer_budget_count": 30,
  "frame_count": 354,
  "frame_rasterizer_count": 354,
  "new_gen_gc_count": 0,
  "old_gen_gc_count": 0,
  "frame_build_times": [

  "average_frame_build_time_millis": 1.7559494382022474,
  "90th_percentile_frame_build_time_millis": 3.62,
  "99th_percentile_frame_build_time_millis": 18.916,
  "worst_frame_build_time_millis": 33.934,
  "missed_frame_build_budget_count": 6,
  "average_frame_rasterizer_time_millis": 9.293120448179277,
  "90th_percentile_frame_rasterizer_time_millis": 12.798,
  "99th_percentile_frame_rasterizer_time_millis": 65.41,
  "worst_frame_rasterizer_time_millis": 119.423,
  "missed_frame_rasterizer_budget_count": 29,
  "frame_count": 356,
  "frame_rasterizer_count": 357,
  "new_gen_gc_count": 0,
  "old_gen_gc_count": 0,
  "frame_build_times": [

  "average_frame_build_time_millis": 1.704057534246574,
  "90th_percentile_frame_build_time_millis": 3.629,
  "99th_percentile_frame_build_time_millis": 18.656,
  "worst_frame_build_time_millis": 38.161,
  "missed_frame_build_budget_count": 6,
  "average_frame_rasterizer_time_millis": 8.447461748633874,
  "90th_percentile_frame_rasterizer_time_millis": 12.113,
  "99th_percentile_frame_rasterizer_time_millis": 30.952,
  "worst_frame_rasterizer_time_millis": 40.757,
  "missed_frame_rasterizer_budget_count": 23,
  "frame_count": 365,
  "frame_rasterizer_count": 366,
  "new_gen_gc_count": 0,
  "old_gen_gc_count": 0,
  "frame_build_times": [
```

## after
```
  "average_frame_build_time_millis": 1.9809107692307677,
  "90th_percentile_frame_build_time_millis": 4.836,
  "99th_percentile_frame_build_time_millis": 17.616,
  "worst_frame_build_time_millis": 34.873,
  "missed_frame_build_budget_count": 5,
  "average_frame_rasterizer_time_millis": 20.8045582822086,
  "90th_percentile_frame_rasterizer_time_millis": 81.98,
  "99th_percentile_frame_rasterizer_time_millis": 112.914,
  "worst_frame_rasterizer_time_millis": 119.172,
  "missed_frame_rasterizer_budget_count": 83,
  "frame_count": 325,
  "frame_rasterizer_count": 326,
  "new_gen_gc_count": 0,
  "old_gen_gc_count": 0,
  "frame_build_times": [

  "average_frame_build_time_millis": 1.6553520000000002,
  "90th_percentile_frame_build_time_millis": 2.537,
  "99th_percentile_frame_build_time_millis": 20.072,
  "worst_frame_build_time_millis": 37.377,
  "missed_frame_build_budget_count": 6,
  "average_frame_rasterizer_time_millis": 8.136740641711222,
  "90th_percentile_frame_rasterizer_time_millis": 11.201,
  "99th_percentile_frame_rasterizer_time_millis": 29.918,
  "worst_frame_rasterizer_time_millis": 42.389,
  "missed_frame_rasterizer_budget_count": 17,
  "frame_count": 375,
  "frame_rasterizer_count": 374,
  "new_gen_gc_count": 0,
  "old_gen_gc_count": 0,
  "frame_build_times": [

  "average_frame_build_time_millis": 1.6299430894308966,
  "90th_percentile_frame_build_time_millis": 2.812,
  "99th_percentile_frame_build_time_millis": 17.016,
  "worst_frame_build_time_millis": 40.751,
  "missed_frame_build_budget_count": 6,
  "average_frame_rasterizer_time_millis": 8.856346883468834,
  "90th_percentile_frame_rasterizer_time_millis": 12.153,
  "99th_percentile_frame_rasterizer_time_millis": 34.18,
  "worst_frame_rasterizer_time_millis": 111.3,
  "missed_frame_rasterizer_budget_count": 27,
  "frame_count": 369,
  "frame_rasterizer_count": 369,
  "new_gen_gc_count": 0,
  "old_gen_gc_count": 0,
  "frame_build_times": [
```

Looking at the median results (8.85ms and 9.29ms) that's a 5% decrease.  Given the wild variance in that test though it's hard to draw conclusions from it.